Implement a basic rolling restart mechanism for any service managed by CDM
The algorithm is this: iterate the list of role types for a service, and then restart the instances of each type in batches, in order. Using the `/roleCommands/restart` endpoint which restarts a batch of role instances and reports on the success, we are able to issue batches of restarts. Each batch must be at least 90% successful for the next one to take place.

For services we care deeply about, we hardcode an order to restart the different kinds of service members in from least-scary to most-scary. My reasoning here is that we want to detect any kind of failure early before restarting the masters if possible. If the config is broken or the processes are systematically not going to come back up for any reason, I’d rather find that out on the doinky worker nodes instead of the coordinators for the services. I can’t think of any situations where we actually want to restart the coordinators first, but we should reason it out more for sure.

I also added a delay of 20s after each batch to wait for that batch of processes to actually finish connecting and coming online. This is arbitrary but I figure it’s good enough, and since a computer is doing this work in the background I am not really worried about it taking too long.

What this lacks: any kind of rack awareness, any kind of parallelism other than submitting batch commands to CDM, any kind of speed, or much production validation.